Create 0-Guessing_game.c
The number-guessing game is a simple game where the user is asked to guess a random number generated by the program. The program gives hints to the user to help them guess the number. The game continues until the user correctly guesses the number. the game is personalized to refer to the user by the username he/she/they enters.